tweaks for new solenoid; pulse scheduling from hall sensor

+++ SW/Board/src/engine_controller.h
@@ -10,7 +10,7 @@ struct EngineController
     volatile Sensor &sensor;
     volatile SignalTracker signal_tracker;
     volatile PID pid;
-    volatile uint8_t turnoff_counter = 0;
+    volatile uint16_t turnoff_counter = 0;
     volatile uint16_t control_counter = 0;
     volatile bool last_signal = false;
     volatile uint8_t last_comp = 0;
